+++ src/index.ts
@@ -2,7 +2,7 @@ import { app } from './initializers/bolt'
 import echo from './commands/echo'
 import cem_register from './commands/cem_register'
 import cem_new from './commands/cem_new'
-import res_cem_new from './actions/res_cem_new'
+import res_cem_new from './listeners/res_cem_new'
 ;(async () => {
   // Start your app
   await app.start(process.env.PORT || 3000)
